fix: fix problem with type width and <<

+++ HW11/include/utils.h
@@ -5,6 +5,6 @@
 
 #include "temperature_record.h"
 
-uint8_t date_to_int(temperature_record_t* record);
+uint64_t date_to_int(temperature_record_t* record);
 
 #endif
